More experimentation with dynamic subclasses.
Continued to flesh out implementation of dynamic subclasses, but appear to have run into a wall where the deserialization infrastructure won't know about the dynamic subclasses.

The deserializers are created for the TServiceMessage ahead of time and, therefore, will only be able to deserialize that exact type. If the deserializer infrastructure was actually changed so that the TServiceMessage was passed in to each Deserialize call then this would not be an issue as I could replace TServiceMessage with TServiceMessageSubclass on the fly when making the deserialization call.

Will revisit this approach down the line, but will likely stick with requiring consumers to subclass a base PeekLockMessage themselves for now.